+++ src/Analysis/mpl_scoretree.h
@@ -12,23 +12,23 @@
 #include "mpl_parsim.h"
 
 typedef void    (*mpl_clipreopt)(mpl_node* src, mpl_node* csite, mpl_tree* t);
-typedef double  (*mpl_scoretry)(const double lim, mpl_node* src, mpl_node* tgt, mpl_tree* t);
+typedef long  (*mpl_scoretry)(const long lim, mpl_node* src, mpl_node* tgt, mpl_tree* t);
 
 MPL_RETURN    mpl_init_parsimony(mpl_matrix* m);
-double  mpl_fullpass_parsimony(mpl_tree* t);
-double  mpl_length_only_parsimony(const double lim, mpl_tree* t);
-double  mpl_fullpass_subtree(mpl_node* subtr, mpl_tree* t);
+long  mpl_fullpass_parsimony(mpl_tree* t);
+long  mpl_length_only_parsimony(const long lim, mpl_tree* t);
+long  mpl_fullpass_subtree(mpl_node* subtr, mpl_tree* t);
 void    mpl_update_src_actives(mpl_node* start, const mpl_node* end);
-double  mpl_fullpass_parsimony_na_only(const double lim, mpl_node* start, mpl_tree* t);
-double  mpl_partpass_parsimony(mpl_node* start, mpl_tree* t);
+long  mpl_fullpass_parsimony_na_only(const long lim, mpl_node* start, mpl_tree* t);
+long  mpl_partpass_parsimony(mpl_node* start, mpl_tree* t);
 void    mpl_src_root_parsimony(mpl_node* src);
-double  mpl_score_try_parsimony(const double, const double lim, mpl_node* src, mpl_node* tgt, mpl_tree* t);
+long  mpl_score_try_parsimony(const long, const long lim, mpl_node* src, mpl_node* tgt, mpl_tree* t);
 void    mpl_scoretree_copy_original_characters(void);
 void    mpl_scoretree_restore_original_characters(mpl_matrix* m, mpl_tree* t);
 void    mpl_tempset_stdtype(void);
 void    mpl_tempreset_natype(void);
 void    mpl_parsim_update_root_prupset(const long left, const long right, const long n, mpl_matrix* m);
-double  mpl_scoretree_calc_abs_minscores(void);
+long  mpl_scoretree_calc_abs_minscores(void);
 void mpl_do_ratchet_weights(void);
 void mpl_reset_std_weights(void);
 #endif /* mpl_scoretree_h */
