+++ src/Characters/mpl_parsim.h
@@ -21,10 +21,10 @@
 typedef struct _matrix mpl_matrix;
 typedef struct mpl_parsdat mpl_parsdat;
 
-typedef double  (*mpl_dnfxn)(const long left, const long right, long n, mpl_parsdat* pd);
+typedef long  (*mpl_dnfxn)(const long left, const long right, long n, mpl_parsdat* pd);
 typedef void    (*mpl_upfxn)(const long left, const long right, long n, long anc, mpl_parsdat* pd);
 typedef void    (*mpl_branchfxn)(const long n, const long anc, mpl_parsdat* pd);
-typedef double  (*mpl_locfxn)(const double lim, const double base, const long src, const long tgt1, const long tgt2, const long troot, mpl_parsdat* pd);
+typedef long  (*mpl_locfxn)(const long lim, const long base, const long src, const long tgt1, const long tgt2, const long troot, mpl_parsdat* pd);
 
 typedef struct mpl_parsdat {
     
@@ -42,16 +42,16 @@ typedef struct mpl_parsdat {
     long*           nndindices;
     long**          ndindexbufs;
     bool            usndidx; // Flag for using node indices
-    double          minscore;
-    double          absminscore; // Absolute minimum score for partition
-    double          cminscore; // A current minimum score based on state of tree and applicable-only changes
-    double          crecall;
+    long          minscore;
+    long          absminscore; // Absolute minimum score for partition
+    long          cminscore; // A current minimum score based on state of tree and applicable-only changes
+    long          crecall;
     long*           minchanges;
     bool            dobreak;
     int             doeschange; // Number of times a character changes
-    double          scorerecall; // For characters needing full-pass estimation during a partial reopt.
-    double          score; // The score for this parsimony type.
-    double          tryscore; // Temporary score for an insertion attempt.
+    long          scorerecall; // For characters needing full-pass estimation during a partial reopt.
+    long          score; // The score for this parsimony type.
+    long          tryscore; // Temporary score for an insertion attempt.
     mpl_charbuf*    cbuf; // The buffer to which this subset refers
     mpl_parsim_t    parstype;
     bool            isNAtype;
@@ -79,88 +79,88 @@ void mpl_parsim_reset_nchanges(mpl_parsdat* pd);
 void mpl_parsim_set_type (const mpl_gap_t gaphandl, const mpl_parsim_t ptype, mpl_parsdat* pd);
 void mpl_parsim_add_data_column_to_buffer (mpl_discr* col, mpl_charinfo* ci, mpl_charbuf* cb, mpl_parsdat* pd);
 void mpl_parsim_setup_tips(mpl_matrix* m, mpl_parsdat* pd);
-double mpl_fitch_downpass (const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_fitch_downpass (const long left, const long right, const long n, mpl_parsdat* pd);
 void mpl_fitch_uppass (const long left, const long right, const long n, const long anc, mpl_parsdat* pd);
 void mpl_fitch_root(const long n, const long anc, mpl_parsdat* pd);
 void mpl_tip_update(const long tipn, const long anc, mpl_parsdat* pd);
 void mpl_na_tip_update(const long tipn, const long anc, mpl_parsdat* pd);
-double mpl_fitch_local_check (const double lim, const double base, const long src, const long tgt1, const long tgt2, const long troot, mpl_parsdat* pd);
-double mpl_fitch_na_tip2tip_check (const double lim, const double base, const long src, const long tgt1, const long tgt2, const long troot, mpl_parsdat* pd);
+long mpl_fitch_local_check (const long lim, const long base, const long src, const long tgt1, const long tgt2, const long troot, mpl_parsdat* pd);
+long mpl_fitch_na_tip2tip_check (const long lim, const long base, const long src, const long tgt1, const long tgt2, const long troot, mpl_parsdat* pd);
 void mpl_fitch_na_root(const long n, const long anc, mpl_parsdat* pd);
-double mpl_do_src_root(const long left, const long right, const long n, mpl_parsdat* pd);
-double mpl_na_do_src_root(const long left, const long right, const long n, mpl_parsdat* pd);
-double mpl_wagner_na_do_src_root(const long left, const long right, const long n, mpl_parsdat* pd);
-double mpl_fitch_na_first_downpass (const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_do_src_root(const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_na_do_src_root(const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_wagner_na_do_src_root(const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_fitch_na_first_downpass (const long left, const long right, const long n, mpl_parsdat* pd);
 void mpl_fitch_na_first_uppass (const long left, const long right, const long n, const long anc, mpl_parsdat* pd);
-double mpl_fitch_na_second_downpass (const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_fitch_na_second_downpass (const long left, const long right, const long n, mpl_parsdat* pd);
 void mpl_fitch_na_root_finalize(const long n, const long anc, mpl_parsdat* pd);
 void mpl_fitch_na_second_uppass (const long left, const long right, const long n, const long anc, mpl_parsdat* pd);
 void mpl_na_tip_finalize(const long tipn, const long anc, mpl_parsdat* pd);
-double mpl_na_only_parsim_first_downpass (const long left, const long right, const long n, mpl_matrix* m);
+long mpl_na_only_parsim_first_downpass (const long left, const long right, const long n, mpl_matrix* m);
 void mpl_na_only_parsim_do_root(const long n, const long anc, mpl_matrix* m);
 int mpl_na_only_parsim_first_uppass (const long left, const long right, const long n, const long anc, mpl_matrix* m);
 void mpl_na_only_parsim_tip_update(const long n, const long anc, mpl_matrix* m);
-double mpl_na_only_parsim_second_downpass (const long left, const long right, const long n, mpl_matrix* m);
-double mpl_wagner_na_down_reroot (const long left, const long right, const long n, mpl_parsdat* pd);
-double mpl_fitch_na_down_reroot (const long left, const long right, const long n, mpl_parsdat* pd);
-double mpl_fitch_na_local_check (const double lim, const double base, const long src, const long tgt1, const long tgt2, const long troot, mpl_parsdat* pd);
+long mpl_na_only_parsim_second_downpass (const long left, const long right, const long n, mpl_matrix* m);
+long mpl_wagner_na_down_reroot (const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_fitch_na_down_reroot (const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_fitch_na_local_check (const long lim, const long base, const long src, const long tgt1, const long tgt2, const long troot, mpl_parsdat* pd);
 void mpl_parsim_reset_scores(mpl_matrix* m);
-double mpl_parsim_get_std_scores(mpl_matrix* m);
-double mpl_parsim_get_na_scores(mpl_matrix* m);
-double mpl_parsim_first_downpass (const long left, const long right, const long n, mpl_matrix* m);
+long mpl_parsim_get_std_scores(mpl_matrix* m);
+long mpl_parsim_get_na_scores(mpl_matrix* m);
+long mpl_parsim_first_downpass (const long left, const long right, const long n, mpl_matrix* m);
 void mpl_parsim_first_uppass (const long left, const long right, const long n, const long anc, mpl_matrix* m);
-double mpl_parsim_second_downpass (const long left, const long right, const long n, mpl_matrix* m);
-double mpl_fitch_na_down_reroot (const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_parsim_second_downpass (const long left, const long right, const long n, mpl_matrix* m);
+long mpl_fitch_na_down_reroot (const long left, const long right, const long n, mpl_parsdat* pd);
 void mpl_parsim_second_uppass (const long left, const long right, const long n, const long anc, mpl_matrix* m);
 void mpl_parsim_tip_finalize(const long n, const long anc, mpl_matrix* m);
 void mpl_parsim_reset_root_state_buffers(const long n, const long anc, mpl_matrix* m);
 void mpl_parsim_do_src_root(const long left, const long right, const long n, mpl_matrix* m);
 void mpl_parsim_do_src_tip(const long n, mpl_matrix* m);
 void mpl_parsim_update_active_sets(const long left, const long right, const long n, mpl_matrix* m);
-double mpl_parsim_local_check (const double lim, const double base, const long src, const long tgt1, const long tgt2, const long troot, mpl_matrix* m);
-double mpl_fitch_na_recalc_first_downpass(const long left, const long right, const long n, mpl_parsdat*  pd);
-double mpl_fitch_na_recalc_second_downpass(const long left, const long right, const long n, mpl_parsdat*  pd);
+long mpl_parsim_local_check (const long lim, const long base, const long src, const long tgt1, const long tgt2, const long troot, mpl_matrix* m);
+long mpl_fitch_na_recalc_first_downpass(const long left, const long right, const long n, mpl_parsdat*  pd);
+long mpl_fitch_na_recalc_second_downpass(const long left, const long right, const long n, mpl_parsdat*  pd);
 void mpl_fitch_na_recalc_root(const long n, const long anc, mpl_parsdat* pd);
 int mpl_fitch_na_recalc_first_uppass(const long left, const long right, const long n, const long anc, mpl_parsdat* pd);
 void mpl_na_recalc_tip_update(const long tipn, const long anc, mpl_parsdat* pd);
 static inline unsigned int mpl_smallest_closed_interval(mpl_discr* res, mpl_discr a, mpl_discr b);
-double mpl_wagner_downpass
+long mpl_wagner_downpass
  (const long left, const long right, const long n, mpl_parsdat* pd);
 void mpl_wagner_uppass
  (const long left, const long right, const long n, const long anc, mpl_parsdat* pd);
-double mpl_wagner_local_check
-(const double lim,
- const double base,
+long mpl_wagner_local_check
+(const long lim,
+ const long base,
  const long src,
  const long tgt1,
  const long tgt2,
  const long troot,
  mpl_parsdat* pd);
-double mpl_wagner_na_second_downpass
+long mpl_wagner_na_second_downpass
  (const long left, const long right, const long n, mpl_parsdat* pd);
 void mpl_wagner_na_second_uppass
  (const long left, const long right, const long n, const long anc, mpl_parsdat* pd);
-double mpl_wagner_na_recalc_second_downpass
+long mpl_wagner_na_recalc_second_downpass
  (const long left, const long right, const long n, mpl_parsdat* pd);
-double mpl_wagner_na_local_check
-(const double lim,
- const double base,
+long mpl_wagner_na_local_check
+(const long lim,
+ const long base,
  const long src,
  const long tgt1,
  const long tgt2,
  const long troot,
  mpl_parsdat* pd);
-double mpl_wagner_src_root(const long left, const long right, const long n, mpl_parsdat* pd);
+long mpl_wagner_src_root(const long left, const long right, const long n, mpl_parsdat* pd);
 
 void mpl_parsim_do_root(const long n, const long anc, mpl_matrix* m);
 void mpl_parsim_finalize_root(const long n, const long anc, mpl_matrix* m);
 void mpl_parsim_tip_update(const long n, const long anc, mpl_matrix* m);
-double mpl_parsim_get_score_recall(mpl_matrix* m);
-double mpl_parsim_get_standard_tryscore(mpl_matrix* m);
+long mpl_parsim_get_score_recall(mpl_matrix* m);
+long mpl_parsim_get_standard_tryscore(mpl_matrix* m);
 void mpl_reset_state_buffs(const long nrows, mpl_parsdat* pd);
 void mpl_parsim_reset_state_buffers(mpl_matrix *m);
-double mpl_parsim_get_na_remaining_minscore(mpl_matrix* m);
-double mpl_parsim_calc_abs_minscore(mpl_matrix* m);
+long mpl_parsim_get_na_remaining_minscore(mpl_matrix* m);
+long mpl_parsim_calc_abs_minscore(mpl_matrix* m);
 int mpl_parsim_check_nas_updated(mpl_matrix* m);
 
 void mpl_parsim_do_ratchet_weights(const int factor, mpl_charbuf* cb);
